refactor(types): split atom schemas into separate files
