--- packages/types/src/page/atoms.d.ts
@@ -1,419 +1,0 @@
-import { z } from "zod";
-import { type PageComponentBase } from "./base";
-export interface ImageComponent extends PageComponentBase {
-    type: "Image";
-    src?: string;
-    alt?: string;
-}
-export declare const imageComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Image">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Image">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Image">;
-    src: z.ZodOptional<z.ZodString>;
-    alt: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface TextComponent extends PageComponentBase {
-    type: "Text";
-    text?: string;
-}
-export declare const textComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Text">;
-    text: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Text">;
-    text: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Text">;
-    text: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface CustomHtmlComponent extends PageComponentBase {
-    type: "CustomHtml";
-    html?: string;
-}
-export declare const customHtmlComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"CustomHtml">;
-    html: z.ZodOptional<z.ZodString>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"CustomHtml">;
-    html: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    href: z.ZodOptional<z.ZodString>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"CustomHtml">;
-    html: z.ZodOptional<z.ZodString>;
-}, z.ZodTypeAny, "passthrough">>;
-export interface ButtonComponent extends PageComponentBase {
-    type: "Button";
-    label?: string;
-    href?: string;
-    variant?: "default" | "outline" | "ghost" | "destructive";
-    size?: "sm" | "md" | "lg";
-}
-export declare const buttonComponentSchema: z.ZodObject<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Button">;
-    label: z.ZodOptional<z.ZodString>;
-    href: z.ZodOptional<z.ZodString>;
-    variant: z.ZodOptional<z.ZodEnum<["default", "outline", "ghost", "destructive"]>>;
-    size: z.ZodOptional<z.ZodEnum<["sm", "md", "lg"]>>;
-}, "passthrough", z.ZodTypeAny, z.objectOutputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Button">;
-    label: z.ZodOptional<z.ZodString>;
-    href: z.ZodOptional<z.ZodString>;
-    variant: z.ZodOptional<z.ZodEnum<["default", "outline", "ghost", "destructive"]>>;
-    size: z.ZodOptional<z.ZodEnum<["sm", "md", "lg"]>>;
-}, z.ZodTypeAny, "passthrough">, z.objectInputType<{
-    id: z.ZodString;
-    width: z.ZodOptional<z.ZodString>;
-    widthDesktop: z.ZodOptional<z.ZodString>;
-    widthTablet: z.ZodOptional<z.ZodString>;
-    widthMobile: z.ZodOptional<z.ZodString>;
-    height: z.ZodOptional<z.ZodString>;
-    heightDesktop: z.ZodOptional<z.ZodString>;
-    heightTablet: z.ZodOptional<z.ZodString>;
-    heightMobile: z.ZodOptional<z.ZodString>;
-    position: z.ZodOptional<z.ZodEnum<["relative", "absolute"]>>;
-    top: z.ZodOptional<z.ZodString>;
-    left: z.ZodOptional<z.ZodString>;
-    margin: z.ZodOptional<z.ZodString>;
-    marginDesktop: z.ZodOptional<z.ZodString>;
-    marginTablet: z.ZodOptional<z.ZodString>;
-    marginMobile: z.ZodOptional<z.ZodString>;
-    padding: z.ZodOptional<z.ZodString>;
-    paddingDesktop: z.ZodOptional<z.ZodString>;
-    paddingTablet: z.ZodOptional<z.ZodString>;
-    paddingMobile: z.ZodOptional<z.ZodString>;
-    minItems: z.ZodOptional<z.ZodNumber>;
-    maxItems: z.ZodOptional<z.ZodNumber>;
-    desktopItems: z.ZodOptional<z.ZodNumber>;
-    tabletItems: z.ZodOptional<z.ZodNumber>;
-    mobileItems: z.ZodOptional<z.ZodNumber>;
-    clickAction: z.ZodOptional<z.ZodEnum<["none", "navigate"]>>;
-    animation: z.ZodOptional<z.ZodEnum<["none", "fade", "slide"]>>;
-} & {
-    type: z.ZodLiteral<"Button">;
-    label: z.ZodOptional<z.ZodString>;
-    href: z.ZodOptional<z.ZodString>;
-    variant: z.ZodOptional<z.ZodEnum<["default", "outline", "ghost", "destructive"]>>;
-    size: z.ZodOptional<z.ZodEnum<["sm", "md", "lg"]>>;
-}, z.ZodTypeAny, "passthrough">>;
