+++ packages/types/src/page/atoms/Image.d.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+import { type PageComponentBase } from "../base";
+
+export interface ImageComponent extends PageComponentBase {
+  type: "Image";
+  src?: string;
+  alt?: string;
+}
+
+export declare const imageComponentSchema: z.ZodType<ImageComponent>;
