+++ packages/types/src/page/atoms/Text.d.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import { type PageComponentBase } from "../base";
+
+export interface TextComponent extends PageComponentBase {
+  type: "Text";
+  text?: string;
+}
+
+export declare const textComponentSchema: z.ZodType<TextComponent>;
