+++ packages/types/src/page/atoms/index.d.ts
@@ -1,0 +1,4 @@
+export * from "./Image";
+export * from "./Text";
+export * from "./CustomHtml";
+export * from "./Button";
